Update index mappings for: project, study, program (resolves #513)
Update Commons Azul index mappings for the ETL output format as
of 10/16/18 for: project, study, program.

Note: The mapping to program must be updated when the program object
becomes available in the ETL output. The current, specific mapping
to program is an interim solution until the program JSON object is available.