Simplistic change to always index the latest bundle version
Currently the DSS doesn't provide an actual delete bundle operation,
and currently if a data set is reloaded, both the previous bundles and
new bundles for the data set still exist.

Reindexing can process bundles in any order, and with the current
simplistic Commons indexing scheme the latest bundle indexed "wins".
Therefore, always process the latest version of the bundle, on the
assumption that "newer is better".